+++ backend/src/role/models/role.model.ts
@@ -1,12 +1,14 @@
 import { Column, DataType, Model, Table } from "sequelize-typescript";
+import { ApiProperty } from "@nestjs/swagger";
 
 interface RoleCreationAttrs {
   value: string;
   description: string;
 }
 
-@Table({ tableName: "roles", createdAt: false, updatedAt: false  })
+@Table({ tableName: "roles", createdAt: false, updatedAt: false })
 export class RoleModel extends Model<RoleModel, RoleCreationAttrs> {
+  @ApiProperty({ example: "5", description: "Unique identifier" })
   @Column({
     type: DataType.INTEGER,
     unique: true,
@@ -15,9 +17,11 @@ export class RoleModel extends Model<RoleModel, RoleCreationAttrs> {
   })
   id: number;
 
+  @ApiProperty({ example: "5", description: "Role's value" })
   @Column({ type: DataType.STRING, unique: true, allowNull: false })
   value: string;
 
+  @ApiProperty({ example: "5", description: "Role's description" })
   @Column({ type: DataType.STRING, allowNull: false })
   description: string;
 }